backend/foodgram/requirements.txt
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==3.2.15
django-colorfield==0.7.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.4.0
flake8==5.0.4
flake8-broken-line==0.5.0
flake8-isort==4.2.0
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
fonttools==4.37.4
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==1.7.0
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
oauthlib==3.2.1
pep8-naming==0.13.2
Pillow==9.2.0
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.3.0
pyflakes==2.5.0
PyJWT==2.5.0
pyphen==0.13.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tinycss2==1.1.1
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
waitress==2.1.2
weasyprint==56.1
webencodings==0.5.1
zipp==3.8.1
zopfli==0.2.1
